Add script to download upstream information for all modules.
Now we can have upstream information for all modules, and not just
modules hosted on ftp.gnome.org.

The new script (download-upstream-versions) uses feedparser.py, that is
put in the repository for convenience.

Document new files in README.

Add some metadata for some modules, to show it's working. More to come.