Puts more things to do in 'osc gnome ta'.
Now, we can know if a package in G:F is not a link, or if it needs a
manual merge, or if its parent does not exist.

create-database detects most of these errors automatically.

runme adds a .is-devel-project file for GNOME:Factory to help
create-database.

osc gnome was updated for all this, of course.