[download] Support google projects with non-standard tarball names
We used to only support google code projects where the name of the
tarballs is the same as the name of the project. We now have a new
option to specify this.

This also enables us to only check the tarballs we're really interested
in (instead of also checking tarballs with other names).