[obs-checkout] Add threads to parallilize the network-related tasks
obs-checkout is mainly slow because of all the network communication
with the build service. By parallilizing this part with threads, we make
it much faster.

A lot of work here is to make all this work fine with threads. Some
important things:
  + we stop using some methods from osc.core that were downloading some
    data from the build service and implement them ourselves. It's easy
    since it's just access to a public http page.
  + urllib2 does not work well with socket timeouts, so we have to
    create a specific thread to close connections that are taking too
    long.

There's a long comment in _checkout_packages_from_project() explaining
the architecture of all this.

There are two new options:

 + --debug to enable some debug output (mainly related to threads, since
   it's not easy to debug what's going wrong).
 + --threads to control how many helper threads are used. "--threads 1"
   means using no thread.