[obs-checkout] Do not queue a devel project if it's the parent too
Sometimes a project uses itself as a devel project for a specific
package. Don't add it to the devel projects, though.

This is arguably debatable, since you can have:
  <devel package="pkg" project="self"/>
for pkg_bis, and no <devel> tag for pkg. In this case, you would really
want to check out the project itself. However, most users of
--devel-projects-of wouldn't expect this behavior.