[obs-checkout] Download packages metadata of a project too
We save it in $projectdir/osc-gnome-meta. This will be useful for
create-database to have a more detailed view of the relationship between
projects (with devel projects).

We use a worker thread when using threads for this since it's a slow
operation. For this, we add a special queue item for each project where
the package is ''.

We first try to download it, but in case of failure, we just use the
previous version available in the cache; this should still be valid in
most cases anyway.