[create-database] Support a fallback list of upstream versions
Marcus Meissner has a script that fetches the latest upstream version
for various packages. The result is at
  http://www.suse.de/~meissner/factory.upstream.lst

This list is less interesting than the information we gather in
download-upstream-versions (since it doesn't have an URL for the
tarball, eg), but it covers more packages. So we can use it as a
fallback source to get the upstream version.

Add a --upstream-versions-fallback command line option to specify where
the file lives.