[osc-gnome] First working version of todoadmin in the xml port
This removes a lot of code, but we also lose detection of some cases
(FIXME in the code).

Thanks to some infrastructure work that was done, todoadmin should be
quite fast, and doesn't require asking costly data from the build
service.

We remove support for --xs for todoadmin, and instead always use this
behavior: it didn't make sense anyway to tell the user that he had
submitted something to the parenet project.