[create-database] Add the notion of toplevel project
Some projects like openSUSE:Factory, Moblin:Factory and
openSUSE:Factory:Contrib are toplevel projects (ie, they have no parent,
regardless of the packages inside are links or not).

We can determine this with the project metadata: if project name and
parent name are the same, then we have a toplevel project.

A direct effect of this is that some errors like not-real-devel or
parent-without-devel should not exist for packages in those projects.