[osc-gnome] Remove potential trailing slash from packages passed as args
If the user is using autocompletion from the shell to get the package
name passed to osc gnome (by relying on the fact that he already has the
package checked out), then the package will have a trailing slash.

It's inconvenient for the user to have to remove it, so we handle this
case specifically.