[osc-gnome] Make it possible to use more than one repo at the same time
It's sometimes interesting to build against 11.1 and Factory, eg.

The code to detect when to change the metadata of the package has also
been tweaked to be more correct (handling <enable/> or enable tags for a
arch without a repository, eg).

For migration, we move gnome_repo to collab_repos directly. collab_repo
was never an option in a released version of the plugin.