[osc-gnome] Make setup/update branch from the devel project
We add a --nodevelproject option to keep the previous behavior of
forcing the branch where it is.

This will make it easier for users to start using the plugin, with a
default configuration of openSUSE:Factory: they will branch packages
from the right place and be able to contribute faster.