[osc-gnome] Autodetect default repository for builds
We were using openSUSE_Factory before, but it  doesn't exist on all
projects. We now have a heuristic to use the first one in the following
repositories when available:

 + snapshot
 + standard
 + openSUSE_Factory/openSUSE-Factory

Then we try to use a repository prefixed with openSUSE (the last one, to
use the last version), and finally, we fallback on the first repository
available in the project.