[server] Make hermes download as much events needed to satisfy all steps
We download as much events needed to satisfy both the mirror step and
the db step (we do a min()). And then, each step only takes the events
that are of interest to it.

Concretely, it means we can avoid downloading again updates with the
mirror if we just relaunch the script to have the database take into
acount some updates (that were not taken before because of a crash, eg).

Also, fix hermes tests.