[server] Fix project configuration inheritance
A devel project is supposed to inherit a few settings from its parent
(an obvious one being the branch). However, before the database is
created, there's no way to know if a project has a parent, and if yes,
which one.

The only point in time where we have this information is when we look at
devel projects of a project to queue their checkout. Therefore, at this
point, we have to save the configuration of the devel project in a cache
file so that it can easily be fetched again.

(We also save the configuration of main projects this way, since there's
no reason to handle them differently.)