[server] Set opensuse mtime to -1 when we don't use the openSUSE config
Switching between using the openSUSE configuratin and not using it
should trigger a full check. We do this by using -1 as the opensuse
mtime when the openSUSE configuration is not in use.

This means that enabling it will mean going from -1 to a real value
(which is a change) and disabling it will go from a real value to -1
(which is a change too).