[server] Keep the xml status in the status file
This lets us be sure whenever we have, or have not, to create the XML
files. It's quite important that the XML files are kept in sync with the
main database. So even if the database is not updated, but they are out
of sync, we generate the XML files.