[server] Only generate XML files that are needed to be generated
To achieve this, we need to know which projects have seen a change, and
then we compute the list of projects that need a rebuild (it's the list
of changed projects + their devel projects + the parent projects of link
packages).

To get the list of changed projects, we use the hermes events.
ObsDb.upstream_changes() has also been modified to return a list of
changed projects.