[server] Compute the hash of a file to know if it's up-to-date
We were not checking it before, and this lead to a situation where an
old version of the file could be on the disk and, if not updated
because of an error during the hermes event, would stay un-updated
forever.

We solve this by computing the hash of the file to know if it's really
up-to-date.

Other solutions were possible:

  - remove the file in case of error. This is error-prone, and the database
    will be incomplete for the specific package for some time. (Now,
    we'll have old data, but we'll have data, and it can be updated
    later).
  - check that the size of the file on disk is the same. This is
    cheaper, but it's easily wrong. And since we're only working on
    small files, computing a hash is fast.
  - save the metadata of the downloaded files somewhere, so we can
    easily compare it later. This is too error-prone.