[server] Do not fail when removing a checkout of a non-existing package
If a package was added, and then removed and we got both events at the
same time, the hermes strip will only keep the "package removed" event,
instead of removing all the events. This exhibits a crash when trying to
remove a non-existing package.

Add the same obvious fix for project code.

Also add a FIXME to the strip code in hermes, although it's not that
important to fix now.