[server] Do not add non-interesting devel projects in database
When the devel project of a package changes, or when looking at the
devel project of a new package, we were always adding the devel project
to the database, assuming it was checked out. But we first need to look
f the configuration makes this project interesting for the database.