collab: Handle httplib.BadStatusLine exception too in build
When we wait for a build to finish, it's not uncommon to get a
httplib.BadStatusLine exception, since the build service is sometimes
not 100% available. Catch this one, so we can continue waiting.

Note that this is the only place where we handle it, because in all the
other places, this is not really critical.