server: Use mtime and name to determine which spec file to use
With _service, we can have more than one .spec file for a source
package, and the right to use is generall the _service one.

So we also look at, in addition to the name, the mtime of the .spec
files to determine which one we should use: we prefer the more recent
.spec file, and a .spec file that has the right name ($srcpackage.spec,
with an optional "_service:" prefix) if possible.