osc-collab: Add collab_do_package_tracking config option
By default, we disable package tracking as we use a flat structure for
checked out projects (we always check out in the current directory)
which can't work with package tracking; the reason for this design
decision is that this is easier for new contributors.

However, some people do want to use package tracking with collab, so add
the collab_do_package_tracking config option.

Based on idea from Dominique Leuenberger <dimstar@opensuse.org>