osc-collab: Fix crash in todo for packages with no upstream version
With comments, we can now list packages that have no upstream version in
todo. The code was assuming that the upstream version was always a valid
string.

We also remove leading (and trailing) spaces in the string we build,
since a leading space will occur if there's no upstream version.