server: Rework hermes reader
To improve performance, the hermes code was changed to remove the page
argument for feeds. There's instead a last_id argument that we can use,
so the reader now users this.

Also, fetch only one rdf feed merging various hermes subscriptions. This
will be faster and use less resources on the hermes server, but this
will also solve an issue we had, where we could potentially "lose" some
events between feeds.

On a configuration side of things, hermes-urls is gone and replaced by
hermes-baseurl and hermes-feeds.