upstream: Improve filtering of interesting files for modules
Until now, libthai was broken because there were some gtk-im-libthai
tarballs with a higher version.

To fix this, we tighten the filtering for interesting files by making
sure that:

 a) the file contains the module name (was already checked)
 b) 1) the basename of the file matches a tarball for this module name
    or
    2) there is a query field in the URI

b)2) is obviously not perfect, but it works surprisingly well as there's
no false positive because of it (and it's required for aqbanking, for
instance).

No regression at this point.