upstream: Pre-cache HTTP URLs
Because several upstreams are listed on the same HTTP URL (we have
nearly 500 HTTP URL, but only 262 unique ones), we lose a lot of time
re-downloading the same page again and again.

Instead, we do a first pass of pre-caching all the unique HTTP URLs, and
re-use that cache later on.

The cache is files on disks, and we use a hash of the URL to identify
each file. We also make sure to forward potential HTTP errors we get,
via .error files.