obs-db: Add to catchup packages for which we had an issue during mirror
If we cannot get the file list, a file, or the metadata of a package,
add it to catchup for next run.

Note that we create a temporary mirror-error file in the status
directory to save the packages with errors, just after the mirror phase.
This allows a crash after the mirror phase without "forgetting" to
mirror a package (otherwise, we won't mirror on next run because the
hermes message has been handled).

At the end, the mirror-error file is appended to the catchup file.

Also, do not consider this package when updating the db or the xml, to
avoid further errors caused by a missing .spec file for instance (since
it was removed because it was supposed to be re-downloaded).