Add script to download pypi versions
The script gets a list of all available packages on pypi and get the
current version for every package.
This is slow because there's currently no way to get a complete list of
packages with the latest version (as it's done for cpan). The script
get's a list of packages and then gets for every package the version
over the pypi API.
Filled a bug upstream (https://bitbucket.org/pypa/pypi/issue/203) to
support some API/static file to get all package versions.