Add download method for github
None of the existing download methods works for a all github projects.
The github.com/proj/repo/downloads url is obsolete and not all projects
do the releases, just tags. The tag naming does not fit into the
project-1.2.3 schema and is not caught by httpls.

A separate method for github makes sense to me, instead of
overcomplicating httpls. This is initial implementation that ab-uses the
generic tarball matching but is adapted to the github urls.